// model/work-session-4s.ts
export interface WorkSession4SBody {
  workSessionId: string;
  dateStart: string;
  timeStart: string;
  productNumber: string;
  lotNumber: string;
  materialNumber: string;
}

export interface WorkSession4S {
  id: string;
  workSessionId: string;
  productNumber: string;
  lotNumber: string;
  materialNumber: string;
  dateStart: string;
  timeStart: string;
  remark: string | null;
  dateComplete: string | null;
  timeComplete: string | null;
  status: number;
}

export interface WorkSession4SResponse {
  workSession4S: WorkSession4S;
}

export interface CompleteWorkSession4S {
  id: string;
  dateComplete: string;
  timeComplete: string;
}

export interface WorkSession4SByWsId {
  id: string;
  workSessionId: string;
  productNumber: string;
  lotNumber: string;
  materialNumber: string;
  dateStart: string;
  timeStart: string;
  remark: string | null;
  dateComplete: string | null;
  timeComplete: string | null;
  status: number;
}

export interface WorkSession4SListResponse {
  workSession4Ss: WorkSession4SByWsId[];
}